Move FeaturesQueryPath to a separate file
